+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/src/Vendoring/Shared/SemanticConventions.cs
@@ -25,8 +25,6 @@ internal static class SemanticConventions
     public const string AttributeEnduserRole = "enduser.role";
     public const string AttributeEnduserScope = "enduser.scope";
 
-    public const string AttributePeerService = "peer.service";
-
     public const string AttributeHttpMethod = "http.method";
     public const string AttributeHttpUrl = "http.url";
     public const string AttributeHttpTarget = "http.target";
@@ -44,7 +42,6 @@ internal static class SemanticConventions
     public const string AttributeHttpResponseContentLength = "http.response_content_length";
     public const string AttributeHttpResponseContentLengthUncompressed = "http.response_content_length_uncompressed";
 
-    public const string AttributeDbSystem = "db.system";
     public const string AttributeDbConnectionString = "db.connection_string";
     public const string AttributeDbUser = "db.user";
     public const string AttributeDbMsSqlInstanceName = "db.mssql.instance_name";
@@ -117,6 +114,11 @@ internal static class SemanticConventions
     public const string AttributeServerPort = "server.port"; // replaces: "net.host.port" (AttributeNetHostPort)
     public const string AttributeUserAgentOriginal = "user_agent.original"; // replaces: http.user_agent (AttributeHttpUserAgent)
 
+    // v1.23.0 Database spans
+    // https://github.com/open-telemetry/semantic-conventions/blob/release/v1.23.x/docs/database/database-spans.md
+    public const string AttributeNetworkPeerAddress = "network.peer.address"; // replaces: "net.peer.ip" (AttributeNetPeerIp)
+    public const string AttributeNetworkPeerPort = "network.peer.port"; // replaces: "net.peer.port" (AttributeNetPeerPort)
+
     // v1.24.0 Messaging spans
     // https://github.com/open-telemetry/semantic-conventions/blob/v1.24.0/docs/messaging/messaging-spans.md
     public const string AttributeMessagingClientId = "messaging.client_id";
@@ -135,6 +137,17 @@ internal static class SemanticConventions
     public const string AttributeMessagingKafkaDestinationPartition = "messaging.kafka.destination.partition";
     public const string AttributeMessagingKafkaMessageKey = "messaging.kafka.message.key";
     public const string AttributeMessagingKafkaMessageOffset = "messaging.kafka.message.offset";
+
+    // New database conventions as of commit:
+    // https://github.com/open-telemetry/semantic-conventions/blob/25f74191d749645fdd5ec42ae661438cf2c1cf51/docs/database/database-spans.md#common-attributes
+    public const string AttributeDbSystem = "db.system";
+    public const string AttributeDbCollectionName = "db.collection.name";
+    public const string AttributeDbNamespace = "db.namespace";
+    public const string AttributeDbOperationName = "db.operation.name";
+    public const string AttributeDbResponseStatusCode = "db.response.status_code";
+    public const string AttributeDbOperationBatchSize = "db.operation.batch.size";
+    public const string AttributeDbQuerySummary = "db.query.summary";
+    public const string AttributeDbQueryText = "db.query.text";
 
 #pragma warning restore CS1591 // Missing XML comment for publicly visible type or member
 }